refactored dtype
